% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif_Pred.R
\name{dna_rna_sequence}
\alias{dna_rna_sequence}
\title{Define a class for DNA/RNA sequences}
\usage{
dna_rna_sequence(sequence)
}
\arguments{
\item{sequence}{A character string representing a DNA/RNA sequence.}
}
\value{
An object of class "dna_rna_sequence".
}
\description{
Define a class for DNA/RNA sequences
}
