% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif_Pred.R
\name{predictMotif.dna_rna_sequence}
\alias{predictMotif.dna_rna_sequence}
\title{Method dispatch for motif prediction}
\usage{
predictMotif.dna_rna_sequence(sequence, motif)
}
\arguments{
\item{sequence}{An object of class "dna_rna_sequence".}

\item{motif}{A character string representing the motif to be predicted.}
}
\description{
Method dispatch for motif prediction
}
