% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif_Pred.R
\name{predictMotif}
\alias{predictMotif}
\title{Predict motif positions in a sequence}
\usage{
predictMotif(sequence, motif)
}
\arguments{
\item{sequence}{A character string representing a DNA/RNA sequence.}

\item{motif}{A character string representing the motif to be predicted.}
}
\value{
Prints the positions of the motif in the sequence.
}
\description{
This function predicts the positions of a given motif within a sequence.
}
